+++ scripts/world-protection-from-evil.js
@@ -1,10 +1,10 @@
-const version = "9.1";
+const version = "10.0";
 
 Hooks.on("midi-qol.preAttackRoll", function (workflow) {
   let targetActor = workflow.targets.first().actor;
-  let effect = targetActor.data.effects.find(i=> i.data.label === "Protection from Good and Evil" || i.data.label === "Protection from Evil and Good");
+  let effect = targetActor.effects.find(i=> i.data.label === "Protection from Good and Evil" || i.data.label === "Protection from Evil and Good");
   if (effect !== undefined) {
-	let attackerType = ["aberration", "celestial", "elemental", "fey", "fiend", "undead"].includes(workflow.actor.data.data.details?.type?.value);
+	let attackerType = ["aberration", "celestial", "elemental", "fey", "fiend", "undead"].includes(workflow.actor.system.details?.type?.value);
 	if(attackerType) {
 		workflow.disadvantage = true;
 	}